% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaclipcc.launchers.R
\name{anomalyMap.ECV.ipcc}
\alias{anomalyMap.ECV.ipcc}
\title{Launch ECV anomaly map description}
\usage{
anomalyMap.ECV.ipcc(project = "CMIP5", experiment = c("rcp45",
  "rcp85"), future.period = c("near", "mid", "long"), variable, season,
  time.res.orig = c("P1D", "P1M"), proj = "EPSG:54030",
  anomaly.type = c("absolute", "relative"))
}
\arguments{
\item{project}{Currently unused. CMIP5 by default.}

\item{experiment}{code of the experiment}

\item{future.period}{future period}

\item{variable}{variable}

\item{season}{season}

\item{time.res.orig}{original temporal resolution of the variable}

\item{proj}{projection string}

\item{anomaly.type}{type of anomaly (absolute or relative)}
}
\value{
A METACLIP object describing the generation of an anomaly map.
}
\description{
Generates the full graph describing an anomaly map of an ECV
}
\author{
J. Bedia
}
