% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaclipcc.launchers.R
\name{deltaMap.ECV.ipcc}
\alias{deltaMap.ECV.ipcc}
\title{Launch ECV delta map description}
\usage{
deltaMap.ECV.ipcc(project = "CMIP5", experiment = "rcp85",
  future.period = "mid", baseline = c(1986, 2005), variable, season,
  time.res.orig = "P1M",
  proj = "+proj=robin +lon_0=0 +x_0=0 +y_0=0 +ellps=WGS84 +datum=WGS84 +units=m +no_defs",
  delta.type = "absolute")
}
\arguments{
\item{project}{Currently unused. CMIP5 by default.}

\item{experiment}{code of the experiment}

\item{future.period}{future period. Current options include the standard AR5 future time slices.}

\item{baseline}{Vector of length two indicating the start/end years of the baseline (historical) period.
Default to the AR5 baseline \code{c(1986,2005)}.}

\item{variable}{variable}

\item{season}{season}

\item{time.res.orig}{original temporal resolution of the variable}

\item{proj}{projection string}

\item{delta.type}{type of delta (absolute or relative)}
}
\value{
A METACLIP object describing the generation of a delta map.
}
\description{
Generates the full graph describing a delta map of an ECV
}
\author{
J. Bedia
}
