% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaclipcc.helpers.R
\name{getVariableVersion}
\alias{getVariableVersion}
\title{Internal for retrieving a Variable version from the dataset master table}
\usage{
getVariableVersion(Dataset.name, variable)
}
\value{
The variable version for the specified dataset, as a character string
}
\description{
Obtain the \code{ds:withVersionTag} data property associated with a CMIP6 \code{ds:Variable}
}
\seealso{
Other lookup-tables: 
\code{\link{metaclipcc.getFuturePeriod}()},
\code{\link{showIPCCdatasets}()},
\code{\link{showIPCCvars}()}
}
\concept{lookup-tables}
\keyword{internal}
