% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaclipcc.Anomaly.R
\name{metaclipcc.AnomalyCalculation}
\alias{metaclipcc.AnomalyCalculation}
\title{Directed metadata graph construction for Anomaly transformations}
\usage{
metaclipcc.AnomalyCalculation(graph, referenceGraph = NULL,
  anomaly.type = c("absolute", "relative"))
}
\arguments{
\item{graph}{An output from a previous \pkg{metaclipR} function containing a list with the i-graph class object containing
the input grid whose anomaly is to be computed, plus the terminal node from which the Anomaly Step will hang}

\item{referenceGraph}{An output from a previous \pkg{metaclipR} function containing a list with the i-graph class object containing the reference Transformation-class object
used as base to compute the climatology, plus the name of its terminal node}

\item{anomaly.type}{CHaracter string. Either \code{"absolute"} (default), or \code{"relative"}, if the anomaly is computed as a
ratio instead of a difference.}
}
\description{
Build a directed metadata graph describing an anomaly Transformation
}
\details{
This function takes as reference the semantics defined in the Data Source and Transformation ontology
defined in the Metaclip Framework (\url{http://www.metaclip.org}).
}
\author{
J. Bedia
}
\concept{transformation}
