% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaclipcc.AnomalyMap.R
\name{metaclipcc.AnomalyMap}
\alias{metaclipcc.AnomalyMap}
\title{Encode metadata of graphical output for anomaly maps}
\usage{
metaclipcc.AnomalyMap(graph, proj, spatial.ref)
}
\arguments{
\item{graph}{input METACLIP graph structure}

\item{proj}{Projection string}

\item{spatial.ref}{Input spatial reference indicating map spatial extent}
}
\description{
Encode metadata of graphical output for anomaly maps
}
\author{
J Bedia
}
