% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaclipcc.BiasCorrection.R
\name{metaclipcc.BiasCorrection}
\alias{metaclipcc.BiasCorrection}
\title{Directed metadata graph construction for bias correction steps}
\usage{
metaclipcc.BiasCorrection(
  graph,
  TrainingGraph,
  ReferenceGraph,
  ReferenceGraphSpatialExtent = NULL,
  ReferenceGraphRectangularGrid = NULL,
  BC.method = "EQM",
  dc.description = "Bias adjustment of the input data"
)
}
\arguments{
\item{graph}{metaclipR output containing the data top be bias-corrected.}

\item{TrainingGraph}{metaclipR output containing the training data (e.g. 20C3M/historical scenario in
climate change applications etc.)}

\item{ReferenceGraph}{metaclipR output containing the reference predictand (typically observations)}

\item{ReferenceGraphSpatialExtent}{Default to \code{NULL} and unused. Otherwise, this points to a SpatialExtent class node
containing the horizontal spatial extent information of the observations. This will update the Spatial extent of the calibrated
dataset to that of the reference observations used for calibration.}

\item{ReferenceGraphRectangularGrid}{Default to \code{NULL} and unused. Otherwise, this points to a ds:RectangularGrid class node
containing the grid definition of the predictand. This will update the Spatial extent of the calibrated
dataset to that of the reference observations used for calibration.}

\item{BC.method}{Character string indicating the name of the bias correction method. Currently accepted
values are \code{"EQM"} and \code{"ISIMIP3"} (the only ones implemented so far in the Atlas Chapter).}

\item{dc.description}{Default to \code{NULL} and unused. Otherwise, this is a character string that will be appendend as a
"dc:description" annotation to the ds:Calibration node.}
}
\description{
Build a directed metadata graph from bias correction routines. A IPCC Atlas targeted version of the
more general \code{\link[metaclipR]{metaclipR.BiasCorrection}}.
}
\details{
This function takes as reference the semantics defined in the Calibration ontology defined in the
Metaclip Framework (\url{http://www.metaclip.org/}). These in turn are partially based on the VALUE Framework (Gutiérrez et al. 2018)
}
\references{
Gutiérrez et al, 2018. An intercomparison of a large ensemble of statistical downscaling methods over Europe:
Results from the VALUE perfect predictor cross-validation experiment. International Journal of Climatology.
https://doi.org/10.1002/joc.5462
}
\author{
J. Bedia
}
