% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaclipcc.ClimateIndex.R
\name{metaclipcc.ClimateIndex}
\alias{metaclipcc.ClimateIndex}
\title{Encode metadata for IPCC product climate index definition}
\usage{
metaclipcc.ClimateIndex(graph, index.code, new.time.res = "P1M")
}
\arguments{
\item{graph}{Input graph}

\item{index.code}{Index code}

\item{new.time.res}{Temporal resolution of the index. Default to monthly}
}
\value{
A metaclipR object
}
\description{
Encode metadata for IPCC product climate index definition
}
\details{
The function asssumes that the index.code passed as argument is an individual instance of the ds:ClimateIndex class,
defined in the ipcc_terms vocabulary
}
\author{
J Bedia
}
