% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaclipcc.ColorPalette.R
\name{metaclipcc.ColorPalette}
\alias{metaclipcc.ColorPalette}
\title{Color Palette}
\usage{
metaclipcc.ColorPalette(input.ecv, diverging)
}
\arguments{
\item{input.ecv}{Input ECV determining the character of the palette.
Essentially, two different schemes will be used for either precipitation or temperature-related variables}

\item{diverging}{Logical. If TRUE, a diverging palette (for deltas). Otherwise a sequential palette is assumed.}
}
\description{
Creates a sinle node graph that defines a go:ColorPalette-class object
}
\author{
J Bedia
}
\keyword{internal}
