% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaclipcc.DatasetSubset.R
\name{metaclipcc.DatasetSubset}
\alias{metaclipcc.DatasetSubset}
\title{Directed metadata graph construction for IPCC Atlas product data sources}
\usage{
metaclipcc.DatasetSubset(
  metaclipcc.Dataset,
  Dataset.name,
  variable,
  time.res.orig = c("P1D", "P1M"),
  ipcc.region = "GlobalExtent",
  season = 1:12,
  years
)
}
\arguments{
\item{metaclipcc.Dataset}{METACLIP description of the Dataset, as returned by \code{\link{metaclipcc.Dataset}}.}

\item{Dataset.name}{Name (label) of the Dataset. This dataset must be included in the internal
lookup table, that can be accessed via \code{\link{showIPCCdatasets}}.}

\item{variable}{Code of the ECV. This must be an individual instance of \code{ds:Variable}.}

\item{time.res.orig}{Temporal resolution of the original data, as downloaded from ESGF. Monthly (\code{"P1M"})for ECV maps,
and daily (\code{"P1D"}) for index calculation.}

\item{ipcc.region}{Code of the IPCC region. This must be an individual instance of \code{ds:HorizontalExtent}.}

\item{season}{Season. Integer vector of (correlative) months, e.g.: \code{c(12,1,2)} for DJF (boreal winter).}

\item{years}{Integer vector of length two with start/end year of the period, e.g.: \code{c(2016,2035)} for the AR5 baseline.}
}
\description{
Build a directed metadata graph describing a subset of Dataset
}
\author{
J. Bedia
}
