% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaclipcc.CCsignal.R
\name{metaclipcc.DeltaChangeCalculation}
\alias{metaclipcc.DeltaChangeCalculation}
\title{Directed metadata graph construction for Climate Change Signal calculation}
\usage{
metaclipcc.DeltaChangeCalculation(graph, referenceGraph = NULL,
  delta.type = c("absolute", "relative"))
}
\arguments{
\item{graph}{An output from a previous \pkg{metaclipR} function containing a list with the i-graph class object containing
the input grid whose anomaly is to be computed, plus the terminal node from which the Anomaly Step will hang}

\item{referenceGraph}{An output from a previous \pkg{metaclipR} function containing a list with the i-graph class object containing the reference Transformation-class object
used as base to compute the climatology, plus the name of its terminal node}

\item{delta.type}{Character string. Either \code{"absolute"} (default), or \code{"relative"}, if the delta change is computed as a
ratio instead of a difference.}
}
\description{
Build a directed metadata graph describing a Climate Change Signal calculation
}
\details{
This function takes as reference the semantics defined in the Data Source and Transformation ontology
defined in the Metaclip Framework (\url{http://www.metaclip.org}).
}
\author{
J. Bedia
}
\concept{transformation}
