% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaclipcc.File.R
\name{metaclipcc.File}
\alias{metaclipcc.File}
\title{Directed metadata graph construction for AR6 Atlas Intermediate product}
\usage{
metaclipcc.File(
  project = "CMIP5",
  variable = NULL,
  climate.index = NULL,
  experiment,
  test.mode = FALSE
)
}
\arguments{
\item{project}{Project. Default to \code{"CMIP6"}.}

\item{variable}{Code of the input ECV (it can be omitted if \code{climate.index} is indicated).
Current accepted values are restricted to \code{"tas", "meanpr", "TX", "TN", "prsn", "wind"}
and \code{"siconc", "ph", "tos"} for oceanic variables (CMIP6 only).}

\item{climate.index}{If the map is for a climate index, name of the index. Otherwise NULL (the default). Currently accepted values are restricted to
the set of indices to be included in the Atlas, namely:
\code{"TXx", "TNn", "Rx1day", "Rx5day", "spi6", "CDD", "tx35", "tx40", "cd", "hdd", "fd"}, as well as the
bias adjusted versions of \code{"tx35isimip", "tx40isimip", "fdisimip"}.}

\item{experiment}{Experiment results displayed in the map. Accepted values are restricted to \code{"historical", "rcp26", "rcp45", "rcp85"},
for CORDEX and CMIP5 products, and \code{"historical", "ssp126", "ssp245", "ssp370", "ssp460" and "ssp585"} for CMIP6 products.}

\item{test.mode}{For internal use only. When the test mode is on, only the first two models are used.}
}
\description{
Build a directed metadata graph describing an intermediate product of the AR6 Interactive Atlas
}
\details{
This function has been built by 'prunning' metaclipcc.Map, to accomodate the provenance description of "intermediate products".
}
\author{
J. Bedia
}
