% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaclipcc.helpers.R
\name{metaclipcc.getFuturePeriod}
\alias{metaclipcc.getFuturePeriod}
\title{Get future period}
\usage{
metaclipcc.getFuturePeriod(project, model, future.period, rcp)
}
\arguments{
\item{project}{Either \code{"CMIP5"} or \code{"CMIP6"}}

\item{model}{GCM name, as encoded in the reference lookup \code{dataset_table.csv}. Type  \code{unique(showIPCCdatasets(names.only = FALSE)$GCM)}}

\item{future.period}{Current options include the standard AR5 future time slices for near term, \code{"2021-2040"},
medium term \code{"2041-2060"} and long term \code{"2081-2100"}, and the global warming levels
of +1.5 degC \code{"1.5"}, +2 \code{"2"} and +3 \code{"3"}.}

\item{rcp}{Experiment. Accepted values are restricted to \code{"historical", "rcp26", "rcp45", "rcp60", "rcp85"}}
}
\value{
An integer vector of length two, formed by start and end year of the future time slice
}
\description{
Get a future period time slice ready to be used by \code{metaclipR} functions
}
\seealso{
Other lookup-tables: 
\code{\link{showIPCCdatasets}()},
\code{\link{showIPCCvars}()}
}
\concept{lookup-tables}
\keyword{internal}
