% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaclipcc.HorizontalExtent.R
\name{metaclipcc.HorizontalExtent}
\alias{metaclipcc.HorizontalExtent}
\title{Set a reference SpatialExtent}
\usage{
metaclipcc.HorizontalExtent(region = NULL, xmin = NULL, xmax = NULL,
  ymin = NULL, ymax = NULL)
}
\arguments{
\item{region}{Character string designating a known region.}

\item{xmin}{xmin}

\item{xmax}{xmax}

\item{ymin}{ymin}

\item{ymax}{ymax}
}
\value{
A metaclipR list (igraph-class structure + terminal node name)
}
\description{
Creates a graph containing a SpatialExtent definition, that can be used later as spatial reference of
subsetting, regridding operations etc. so all steps point to the same reference spatial extent
}
\author{
J Bedia
}
