% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaclipcc.MapProjection.R
\name{metaclipcc.MapProjection}
\alias{metaclipcc.MapProjection}
\title{Map Projection}
\usage{
metaclipcc.MapProjection(proj)
}
\arguments{
\item{proj}{A Valid individual instance of class go:MapProjection}
}
\description{
Creates a sinle node graph that defines a map projection
}
\author{
J Bedia
}
\keyword{internal}
