% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaclipcc.Regridding.R
\name{metaclipcc.Regridding}
\alias{metaclipcc.Regridding}
\title{Directed metadata graph construction for Regridding Transformations}
\usage{
metaclipcc.Regridding(graph, RefSpatialExtent = NULL,
  InterpolationMethod, dc.description = NULL)
}
\arguments{
\item{graph}{Input metaclip graph structure}

\item{RefSpatialExtent}{A reference spatial extent used for interpolation. The reference spatial extent can be initiated with \code{\link{metaclipR.SpatialExtent}}}

\item{InterpolationMethod}{Interpolation method. Current possible choices include \code{"nearest"}, \code{"bilinear"}, \code{"bicubic"},
\code{"IDW"}, \code{"spline"} and \code{"ConservativeRemapping"}, although some of them won't be probably used in the IPCC AR6 Atlas}

\item{dc.description}{Default to \code{NULL} and unused. Otherwise, this is a character string that will be appendend as a
"dc:description" annotation to the ds:Interpolation-class node.}
}
\description{
Build a directed metadata graph describing a regridding step
}
\author{
J. Bedia
}
\keyword{internal}
