% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaclipcc.helpers.R
\name{showIPCCvars}
\alias{showIPCCvars}
\title{Table of IPCC Atlas Variables}
\usage{
showIPCCvars(names.only = TRUE)
}
\arguments{
\item{names.only}{Logical flag indicating that only the variable names are shown. Default to \code{TRUE}. Otherwise, the full metadata table is shown}
}
\value{
Either a vector with all dataset names (when \code{names.only = TRUE}) or a \code{data.frame} with all the associated metadata.
}
\description{
Shows the internal table with IPCC Variables and relevant metadata
}
\examples{
showIPCCvars()
}
\seealso{
Other lookup-tables: \code{\link{metaclipcc.getFuturePeriod}},
  \code{\link{showIPCCdatasets}}
}
\author{
J Bedia
}
\concept{lookup-tables}
\keyword{internal}
