% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaclipcc.helpers.R
\name{showOBSdatasets}
\alias{showOBSdatasets}
\title{Table of Observational reference datasets}
\usage{
showOBSdatasets(names.only = TRUE)
}
\arguments{
\item{names.only}{Logical flag indicating that only the internal dataset names are shown. Default to \code{TRUE}.
Otherwise, the full metadata table is shown}
}
\value{
Either a vector with all dataset names (when \code{names.only = TRUE}) or a \code{data.frame}
with all the associated metadata.
}
\description{
Shows the internal table with Observtaional Datasets and relevant metadata
}
\details{
All the observational datasets used as reference should be described as
individual instances of the ObservationalDataset Class in the datasource vocabulary.
}
\examples{
showOBSdatasets()
}
\seealso{
Other lookup-tables: \code{\link{showIPCCdatasets}},
  \code{\link{showIPCCvars}}
}
\author{
J Bedia
}
\concept{lookup-tables}
\keyword{internal}
