% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{update_dataset_table}
\alias{update_dataset_table}
\title{Update reference dataset table}
\usage{
update_dataset_table(
  projects = c("CMIP5", "CMIP6", "CMIP6Omon"),
  out.file = "/tmp/dataset_table.csv"
)
}
\arguments{
\item{projects}{projects vector string (several allowed). Currently just \code{"CMIP5"} and \code{"CMIP6"}.}

\item{out.file}{Full path to output file table. Default to "/tmp/dataset_table.csv". See Details.}
}
\description{
Update var/index availability in master table from a curated reference JSON
}
\details{
The function will replace the appropriate fields within a copy of the master table dataset_table.csv
in the inst folder, and store it in the target directory (writing permission is assumed).

Note that the dataset_table.csv in the inst folder will have to be replaced manually afterwards, and the package
compiled for a new updated version.

The reference JSON storing the curated version of the IA ensemble members is at
\url{https://gitlab.predictia.es/ipcc/data-homogenizers/-/blob/master/total_members_curated.json}.
An up-to-date copy of this file is stored in 'inst/total_members_curated.json'.
}
\author{
J. Bedia
}
\keyword{internal}
