% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaclipcc.ClimateIndex.R
\name{metaclipcc.ClimateIndex}
\alias{metaclipcc.ClimateIndex}
\title{Encode metadata for IPCC product climate index definition}
\usage{
metaclipcc.ClimateIndex(graph.list, index.code)
}
\arguments{
\item{graph.list}{A list of input graphs.}

\item{index.code}{Index code}
}
\value{
A metaclipR object
}
\description{
Encode metadata for IPCC product climate index definition
}
\details{
The function asssumes that the index.code passed as argument is an individual instance of the ds:ClimateIndex class,
defined in the ipcc_terms vocabulary
}
\author{
J Bedia
}
