% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaclipcc.Dataset.R
\name{metaclipcc.Dataset}
\alias{metaclipcc.Dataset}
\title{Directed metadata graph construction for IPCC Atalas product data sources}
\usage{
metaclipcc.Dataset(Dataset.name = NULL)
}
\arguments{
\item{Dataset.name}{Name (label) of the Dataset. This dataset must be included in the internal
lookup table, that can be accessed via \code{\link{showIPCCdatasets}}.}
}
\description{
Build a directed metadata graph describing a data source. This is usually the initial
step to build METACLIP graphs
}
\author{
J. Bedia
}
