% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaclipcc.DeltaMap.R
\name{metaclipcc.DeltaMap}
\alias{metaclipcc.DeltaMap}
\title{Encode metadata of graphical output for delta maps}
\usage{
metaclipcc.DeltaMap(graph, proj, spatial.ref)
}
\arguments{
\item{graph}{input METACLIP graph structure}

\item{proj}{Projection string}

\item{spatial.ref}{Input spatial reference indicating map spatial extent}
}
\description{
Encode metadata of graphical output for delta maps
}
\author{
J Bedia
}
