% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaclipcc.Map.R
\name{metaclipcc.Map}
\alias{metaclipcc.Map}
\title{Directed metadata graph construction for AR6 Atlas Map products}
\usage{
metaclipcc.Map(project = "CMIP5", variable = NULL,
  climate.index = NULL, delta = NULL, experiment, baseline,
  future.period = NULL, season, bias.adj.method = NULL,
  ref.obs.dataset = NULL, proj, map.bbox = NULL, test.mode = FALSE)
}
\arguments{
\item{project}{Project. Unused so far. Default to \code{"CMIP5"}.}

\item{variable}{Code of the input ECV (it can be omitted if \code{climate.index} is indicated).
Current accepted values are restricted to \code{"tas", "pr", "tasmax", "tasmin"}.}

\item{climate.index}{If the map is for a climate index, name of the index Otherwise NULL (the default). Currently accepted values are restricted to
the set of indices to be included in the Atlas, namely:
\code{"TXx", "TNn", "Rx1day", "Rx5day", "DS", "SPI6", "SPI12", "SPEI6", "SPEI12", "DF6", "DF12", "LFFP", "GDD","T21.5", "TX35", "TX40", "HDD", "FD"}}

\item{delta}{Type of delta map displayed. This can be either \code{"absolute"} or \code{"relative"}. Default to \code{NULL} meaning that the map is not a delta
but an original magnitude.}

\item{experiment}{Experiment results displayed in the map. Accepted values are restricted to \code{"historical", "rcp26", "rcp45", "rcp85"}}

\item{baseline}{Character string indicating the \code{"start-end"} years of the baseline (historical) period. Accepted values are:
\code{"1980-2010"} (WMO standard period), \code{"1986-2005"} (AR5 period) or \code{"1995-2014"} (AR6 period). Internally, there is a tricky part here, see Details.}

\item{future.period}{future period. Default to \code{NULL}, for historical maps (i.e., period defined by the \code{baseline} argument). Otherwise, a character string
indicating either the \code{"start-end"} years of the future period or a (GCM-specific) warming level. Current options include the standard AR5 future time slices
for near term, \code{"2021-2040"}, medium term \code{"2041-2060"} and long term \code{"2081-2100"}, and the global
warming levels of +1.5 degC \code{"1.5"}, +2 \code{"2"} and +3 \code{"3"}.}

\item{season}{season. Integer vector of correlative months, in ascending order, encompassing the target season.}

\item{bias.adj.method}{Default to \code{NULL} and unused. If the map displays a bias-corrected product, a character string idetifying the method. Current accepted values a re \code{"EQM"},
for the standard VALUE empirical quantile mapping method.}

\item{ref.obs.dataset}{Default to \code{NULL}, and unused unless a \code{bias.adj.method} has been specified.
This is the reference observational dataset to perform the correction. This is an individual that must be defined in the datasource vocabulary,
 belonging to either classes \code{ds:ObservationalDataset} or \code{ds:Reanalysis}. Currently accepted values are \code{"EWEMBI"}.}

\item{proj}{Map projection string. Accepted values are \code{"Robin"}, \code{"Arctic"} and \code{"Antarctic"},
for Robinson and WGS84 Arctic/Antarctic Polar stereographic projections.}

\item{map.bbox}{Optional. numeric vector of length 4, containing, in this order the \code{c(xmin, ymin, xmax, ymax)} coordinates of the target area
zoomed in by the user. If missing, then the HorizontalExtent associated to the \code{project} is assumed.}

\item{test.mode}{For internal use only. When the test mode is on, only the first two models are used.}
}
\description{
Build a directed metadata graph describing a Map product of the AR6 Interactive Atlas
}
\details{
\strong{Baseline period definition}

Two of the baseline periods considered (WMO, 1981-2010 and AR6, 1995-2014) go beyond the temporal extent of the historical experiment simulations in AR5, ending in 2005.
In this case, the strategy followed in the different IPCC reports is to fill the missing period between 2006 and the end of the baseline with the years of the future
simulation used in each case. For example, to compute a RCP 8.5 delta using the WMO baseline, the baseline data for the period 2006-2010 will be extracted from the RCP85
simulation, and then binded to the 1981-2005 period of the historical simulation in order to complete the baseline period.
}
\author{
J. Bedia
}
